Convert to use a single root "child" command
Move to a git-like model with a single child command/binary that has subcommands for each of the possible actions. Add help for each action such that `child <subcommand> -h` is always valid and will never actually run anything.

Update aliases to use new main command. Update the README for how to "install" and use the program.