Allow running arc diff even if not rebased
Instead of requiring the branch to be rebased on its parent to diff, just resolve any partial rebase state to determine the base then diff against the base (rather than against the parent branch name). Improve handling of when we resolve the intermediate, in-progress rebase state which should also help smooth over a few other interactions (such as landing after a rebase that had conflicts).